[refacto-gameplay] Command wheel for units

+++ Assets/Scripts/Entities/Interfaces/IDamageable.cs
@@ -1,6 +1,6 @@
 ﻿
 public interface IDamageable
 {
-    void AddDamage(int amount);
+    void AddDamage(float amount);
     void Destroy();
 }
